build graphql client
