+++ app/constants.js
@@ -1,0 +1,2 @@
+export const GRAPHQL_ENDPOINT_LINK = 'http://localhost:4000/'
+export const GRAPQL_ENDPOINT_WEBSOCKET = ''
